Slightly fade out the logos on the homepage
Before they would be the first thing that would hit your eye, which is
suboptimal as we've got a great tagline AND a live editor

Fixed that by making them `opacity: 0.6`, but I'm open to other
solutions. Making them white would be kinda awesome maybe? Any clues how
to do that?