Escape closing style tag in SSR to prevent XSS
When using untrusted input to create styled components
nothing bad can happen client-side because the content
of the inline style is set using DOM methods. However,
when using SSR the values of all `rule.cssText` are
simply concatenated, which is unsafe. Similar to how
closing script tags need to be escaped for initial SSR
state we need to escape style as well. See https://medium.com/node-security/the-most-common-xss-vulnerability-in-react-js-applications-2bdffbcc1fa0