fix removal of PropTypes import in production builds
Frustrating, but because rollup doesn't have an internal concept
of prod vs dev, it doesn't set NODE_ENV accordingly and babel
uses the same config for everything. So we are able to eliminate
the references to prop types, but the module import itself is not
being eliminated without this change.