[v5 candidate] transient props
This commit introduces a new JSX concept called "transient props". When
a prop is prefixed with a $ (e.g. $foo), it will only be applied to
the immediate component and not forwarded to any deeper children or
layers of the component stack.

In the case of styled-components, this equates to a styling-only prop
that you may consume in your style interpolations but will never end
up on the DOM node or in your wrapped component. Other props can still
be accessed normally.