Fix styled components causing update problem
In certain situations styled components react component seems to be
updating when it shouldn't this is a regression from v3 which does don't
have this behaviour. Since PureComponent is too new this patch implements
the basic behaviour of it. This fixes the error we are getting which is in
our big app a deeply nested SurveyJs inputs aren't being updated because
before we can update them we have rendered a new version.

Fixes #2195