shard cGS by runtime instance
fixes #2769

cGS has a different injection profile than typical components, in that
there is a single component ID for the global style and any number of
instances. This creates problems when rendering multiple of the same
cGS with different props, as it is a defacto singleton but we don't
enforce a single cGS instance at once...

Utilizing a simple sharding strategy will create duplicate styles, but
ensure that when instances are removed from the page they're taken out
of the stylesheet in the proper order such that previous still-live
instances will continue to function and not be removed as well.