Normalizes how & are expanded on rulesets with +1 props
So the current behavior of the library, when generating the selectors
for a comma separated selector like so:

```jsx
const Outer = styled.div``
const Inner = styled.div`
  ${Outer}:active &,
  ${Outer}:focus & {
    color: blue;
  }
`

render(<Outer><Inner /></Outer>)
```

Would yield the following CSS selector:

```css
.sc-a:active .sc-b, .sc-a:focus .b { color: blue; }
```

Notice that on the first part of the selector, the "stable" class name was
used to target the inner component (`.sc-b`) whereas the second part of
the selector uses the "instance dependent" class name (`.b`).

According to this documentation change (which at the time of this writing
isn't merged yet):

https://github.com/styled-components/styled-components-website/pull/743

The idea is to favour the "stable" class name in this type of scenario.

The root of the problem had to do with the fact that for some reason the
stylis utility was only applying the "self-reference replacement
magic/logic" to the first `prop` item returned by the pre-processor.

For comma separated rules the preprocessor yields a rule set with as
many `props` as the number of CSS rules that were provided as input:

https://github.com/thysultan/stylis.js/tree/5c0f7c77d30bc3a67c387dfa4b7a09dbf659a57b#abstract-syntax-structure

So the proposed patch is to apply the "self-reference replacement" logic
to all rules in the ruleset (not just the first one).