feat: use rules when generating component id
This commit fixes the "_Warning: Prop className did not match_" error emitted during React DOM hydration. The reason for this was the the internal tracker for component identifiers was incremented for every component construction during server-side compilation, but client-side it happened only for the components in the current view, since JS code is typically code-splitted for the browser.

The fix is to take a different approach for generating component ids entirely. Instead of only looking at the component's displayName, its parent displayName and the running increment, the actual style rule set is also used for generating the hash part of the id. The running increment is still there but only used for components with exactly the same hash. This happens with identical styles and display names. When this collision happens, the generated class name is incremented with a separate literal counter like sc-hash, sc-hash-1, ...