fix(stylis): ensure rules are split coming out of stylis (#4021)
* fix(stylis): ensure rules are split coming out of stylis

When vendor prefixing is enabled, the default behavior in stylis
is for them to come out as a clump. Unfortunately, this cases issues
with CSSOM because some browsers reject rules containing styles
with unknown selectors, e.g. Chrome might reject a -moz prefix or
vice versa.

* chore: undo change, add further failing tests

* chore: prep for failing test later

* fix: partial resolution of stylis rule-splitting issue for prefixing mode

Opened https://github.com/thysultan/stylis/pull/315 to start resolving the
second portion of this work (when media queries are involved.)