Rename LICENSE to LICENSE
# Contributing to Sheetsee.js

This repository (github/jlord/sheetsee.js) is primarily for documentation and the documentation website. It also contains a copy of a full sheetsee.js (one that includes the map, charts and table portions).

### Use this repository to file issues and pull requests on documentation/documentation site.

The documentation is contained in the `/docs` and `/demos` folders in the root. I build the site (on the `gh-pages` branch) from these files so **submit pull requests against the files in `/docs` and `/demos`** and not the `/site` folder or the `gh-pages` branch.

### For issues and pull requests on the JavaScript, use the repository for the affected portion:

- [sheetsee](http://www.github.com/jlord/sheetsee/issues/new)
- [sheetsee-core](http://www.github.com/jlord/sheetsee-core/issues/new)
- [sheetsee-tables](http://www.github.com/jlord/sheetsee-tables/issues/new)
- [sheetsee-maps](http://www.github.com/jlord/sheetsee-maps/issues/new)
- [sheetsee-charts](http://www.github.com/jlord/sheetsee-charts/issues/new)

Thank you much! :heart: :octocat: